Remove custom logging cruft
Previously there was a mixture Plexus logging (used when inside maven)
and of home-made console logger (used when running as a tool, outsides
of Maven).

With this commit custom logger is removed completely and code is
migrated to rely exclusively on Plexus logging.  This means that tools
will have to use Plexus and customize the logger if needed.