Add limited support for absolute JPP artifact files
Artifact files specified as absolute paths pointing to arbitrary
locations are now allowed.  They are implemented as symbolic links
which are pointing to the primary artifact file, which must be
explicitly specified as non-absolute file.

Appropriate suffixes are added depending on artifact version,
classifier and extension.  As absolute symlinks can be placed at any
location, configured repositories are not taken into account and flat
repository layout is always assumed.

Absolute symlinks are only created for artifact files and attached
artifacts, but not for POM files.  Absolute files for artifacts with
no files (i.e. POM artifacts) are silently ignored.