Fix compression size mismatch in installation
When injecting manifest, we're copying all the entries to the modified
file. Apparently, the ZipEntry object contains compressed size which in
rare cases might differ from the size to which the object recompresses
and that may cause an exception. Resetting the compressed size to the
default value fixes the problem for me.