Don't hardcode IT directory layout
Thanks to SureFire configuration propagation added in m2e-core 2.6.2
(part of Eclipse 2024-09 release) it is now possible to stop
hardcoding directory layout and make it configurable via
xmvn-it.properties, which are populated by Maven.

Only IT working directrory (the CWD in which ITs are executed) needs
to be unique between different parallel IT runs, therefore it is
configured via SureFire <systemPropertyVariables> instead of
xmvn-it.properties, which must be the same for all IT forks.

There is one caveat regarding the working directory.  Maven SureFire
creates it, but Eclipse requires it exist before launching ITs.
Therefore maven-resources-plugin is used to create the directory by
copying a dummy empty.txt file into the working directory.