Remove obsolete XMvn Javadoc MOJO
The XMvn Javadoc MOJO was originally created as a lightweight
alternative to the Maven Javadoc Plugin, which at the time still
relied on Maven 2 APIs and introduced many unwanted
dependencies. Since then, the dependencies of the Maven Javadoc Plugin
have been improved, making the XMvn Javadoc MOJO redundant. Therefore,
it is no longer needed and can be removed.